requirements.txt
#requirements
PyYAML
paho-mqtt

